requirements.txt
motor
python-dotenv
tgcrypto
kurigram
openai